rename ota_metadata.util to ota_metadata.utils
